Add wait logic for dataset_upload integration tests. (#182)
Currently, you get a 403 if you try uploading a new version while the
FIRST version is still being processed. After the first version is
processed, you can upload multiple versions in parallel.

The test was flaky because sometimes, the processing hadn't finished
before the new version was being uploaded.

This is a behavior we want to fix. See: b/379171781

In the meantime, let's add some waiting logic.

http://b/362736925