Fix auto-compressed dataset downloads (#194)
Notes and other fixes:
- Previously, we needed to unzip the file in the unit test because it
was actually fetching the archive. This is because our
`DatasetHttpResolver` doesn't put the filename in the path (it passes it
in a query string). Rather than change the implementation in the
resolver, I moved filename handling to the shared stubbed route. cc
@jeward414
- Added mocking for downloads to get 302 redirected (as it does in prod
to GCS). This was needed to get `response.url` to be accurate in
`clients.py`, which drives the auto-decompression feature.
- Added other files to support the auto-(de)compression logic.
- In the interest of keeping this PR small, a follow up PR will add
tests for the Models and Competitions HTTP tests. That may necessitate
moving some routes/etc. around to be shared beyond Datasets.

Fixes #187
http://b/379756505