Add `load_dataset` implementation (#192)
This adds the `load_dataset` implementation without including it to the
exported `kagglehub` module yet. Before adding to the `kagglehub` export
(and making the feature more easily discoverable), I want to add more
testing around all the different file types and settings that are
possible to load. This will come on a follow up PR along with added
documentation on how to use it.

Notes:
- This PR also includes adding the `hf-datasets`/`pandas-datasets`
optional dependencies and hatch environments to install them
- The implementation for this feature has changed a few times. Those
other iterations should be visible in the commit history, but as a
recap:
- `apache_beam` doesn't play nicely with `datasets` due to an underlying
dependency issue
- `mlcroissant` (and Croissant itself) has some major file type
limitations
  - Curated/validated arguments for `datasets` and `pandas`
- **Current:** Simple passthrough `kwargs` for the underlying
dependencies
- As implemented, memory is the main constraint.
[dask](https://docs.dask.org/en/stable/dataframe.html) looks like a
great way to unlock more/larger datasets, but I'm leaning toward leaving
that as an improvement if users run into the in-memory problems. Open to
feedback on this decision (or any others big/small re: the
implementation).

I went a little wide with reviews, but this is feeling more like a
direction I could see us going and want to make sure I'm not missing any
blind spots. This is my first major python contribution, so don't be shy
about feedback 🙏

http://b/379756505