Ask user confirmation before import in `package_import`
Importing a python package involves running (some of) the code defined
within it, and running untrusted code on an insecure environment can be
dangerous. Let's warn the user and get explicit confirmation in such
cases. We skip the checks within Kaggle or Colab Notebooks, when the
logged-in user is the author, and for nested Kaggle Packages.

http://b/393606413