Change colab environment detection logic.
Now that Kaggle uses the Colab as a base image, the `COLAB_RELEASE_TAG`
is set in the image. If users use kagglehub with the Kaggle image, this
value is wrong.

Rely on the ipython kernel name instead and save that value to avoid
having to recompute it each time `is_in_colab_notebook` is called.

I also fixed `get_colab_credentials` to ensure it doesn't fail when used
in the wrong environment and simply skip (same as if the credentials
were not found).

BUG=b/399118140