Change colab environment detection logic. (#230)
Now that Kaggle uses the Colab as a base image, the `COLAB_RELEASE_TAG`
environment variable is set in the Kaggle image which is not a good
signal to detect whether we are running in the Colab environment or not
and this has caused the following issue: https://b/399118140.

Updated the Colab environment detection logic to rely on the ipython
kernel name instead and save that value to avoid having to recompute it
each time `is_in_colab_notebook` is called.

I also fixed `get_colab_credentials` to ensure it doesn't fail when used
in the wrong environment and simply skip (same as if the credentials
were not found). This wasn't necessary given this is gated by
`is_in_colab_notebook()` but it adds extra safety.

BUG=b/399118140