Fix multithreading bug with Kaggle Packages (#248)
We use a ContextVar to capture the current `PackageScope`, but each
scope was using a static stack of tokens to support re-establishing the
previous scope when exiting the currently applied scope, and this ran
into issues when using multithreading. Fix by using a second ContextVar
which is scoped to each thread, so that each thread gets their own token
stack.

First commit adds a test which fails with the old impl, then second
commit applies the fix which results in a successful test run.

BUG=417707383