Fix broken cicd that has new mypy deps. (#256)
Context:
[mm](https://chat.kaggle.net/kaggle/pl/fcxzntt8efdijjp8kc8uyuoz6a)

**Root cause**
There seem to be 2 failures for cicd:
1. Updated `ruff` and other checking tools for non-top level imports
(fixed with inline comments.)
2. Our lint script `mypy --install-types --non-interactive src/kagglehub
tests` will need to install a new dep `scipy-stubs`, which is NOT
available for Python 3.9. This dependence is likely to be introduced in
[this PR](https://github.com/Kaggle/kagglehub/pull/196)
  - Long term solution: retire Python 3.9 since it EOL is 2025/Oct.
- Short term: added exceptions based on [this
doc](https://mypy.readthedocs.io/en/stable/running_mypy.html#missing-imports)

I propose we go for the short term solution to unblock, and keep track
the long term solution in other tasks.